requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-magic-bin>=0.4.14